+++ NEWS.md
@@ -8,6 +8,7 @@ This is a maintenance release due to the changes in the upstream package.
 * Improved the speed performance for `new_data()`/`append_row()`.
 * Improved the speed performance for `as.ts.tbl_ts()` for converting tsibble to `ts`.
 * Added the support for `POSIXlt` as index.
+* When the interval cannot be obtained due to the mismatched index class, an error is issued instead of a warning before.
 
 # tsibble 0.8.3